requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
jinja2
